Revert "Triumph of integer math"
This reverts commit 62f99417d79fea5e4083fc5264db11b40409185f.

+++ src/Units/AWBWUnits.java
@@ -141,7 +141,7 @@ public class AWBWUnits extends UnitModelScheme
     }
 
     @Override
-    public int getDamageRedirect(WeaponModel wm)
+    public double getDamageRedirect(WeaponModel wm)
     {
       return wm.getDamage(this);
     }
